Initial Push
The initial push for the game content. I'm in the midst of working on the tutorial, which will be in its own scene. I might work on things in the GDW project and then copy the content here, which will of course be shared with my other teammates. However, anything specific to the tutorial side will be kept separate.