Step System Basics Complete + Game Models
The step system is essentially complete. All I need now to add button prompts to explain things in greater detail, and have more steps to go through. I also need to change the ending scene.

I added in a bunch of game assets. I don't know if we'll have time to put them in for the tutorial, but they're there.